+++ autcli/commands/account.py
@@ -2,8 +2,14 @@
 The `account` command group.
 """
 
-from autcli.options import keyfile_and_password_options, from_option
-from autcli.options import rpc_endpoint_option, newton_or_token_option, keyfile_option
+from autcli.options import (
+    keyfile_and_password_options,
+    from_option,
+    rpc_endpoint_option,
+    newton_or_token_option,
+    keyfile_option,
+    keystore_option,
+)
 from click import group, command, option, argument, ClickException, Path
 from typing import Dict, List, Optional
 
@@ -23,11 +29,7 @@ def account_group() -> None:
 
 @command(name="list")
 @option("--with-files", is_flag=True, help="also show keyfile names.")
-@option(
-    "--keystore",
-    type=Path(exists=True),
-    help="keystore directory (falls back to config file or ~/.autonity/keystore).",
-)
+@keystore_option()
 def list_cmd(keystore: Optional[str], with_files: bool) -> None:
     """
     List the accounts for files in the keystore directory.
@@ -58,7 +60,7 @@ account_group.add_command(list_cmd)
 @argument("accounts", nargs=-1)
 def info(
     rpc_endpoint: Optional[str],
-    key_file: Optional[str],
+    keyfile: Optional[str],
     accounts: List[str],
     asof: Optional[str],
 ) -> None:
@@ -77,7 +79,7 @@ def info(
     from web3 import Web3
 
     if len(accounts) == 0:
-        account = from_address_from_argument_optional(None, key_file)
+        account = from_address_from_argument_optional(None, keyfile)
         if not account:
             raise ClickException("No accounts specified")
         accounts = [account]
@@ -100,7 +102,7 @@ account_group.add_command(info)
 def balance(
     rpc_endpoint: Optional[str],
     account_str: Optional[str],
-    key_file: Optional[str],
+    keyfile: Optional[str],
     ntn: bool,
     token: Optional[str],
 ) -> None:
@@ -122,7 +124,7 @@ def balance(
         format_newton_quantity,
     )
 
-    account_addr = from_address_from_argument_optional(account_str, key_file)
+    account_addr = from_address_from_argument_optional(account_str, keyfile)
     if not account_addr:
         raise ClickException(
             "could not determine account address from argument or keyfile"
@@ -157,7 +159,7 @@ account_group.add_command(balance)
 @keyfile_option()
 @argument("account_str", metavar="ACCOUNT", default="")
 def lntn_balances(
-    rpc_endpoint: Optional[str], account_str: Optional[str], key_file: Optional[str]
+    rpc_endpoint: Optional[str], account_str: Optional[str], keyfile: Optional[str]
 ) -> None:
     """
     Print the current balance of the given account.
@@ -174,7 +176,7 @@ def lntn_balances(
     from autonity.erc20 import ERC20
     from autonity.utils.denominations import format_newton_quantity
 
-    account_addr = from_address_from_argument_optional(account_str, key_file)
+    account_addr = from_address_from_argument_optional(account_str, keyfile)
     if not account_addr:
         raise ClickException(
             "could not determine account address from argument or keyfile"
@@ -200,7 +202,8 @@ account_group.add_command(lntn_balances)
 
 
 @command()
-@keyfile_option(required=True, output=True)
+@keystore_option()
+@keyfile_option(required=False, output=True)
 @option(
     "--extra-entropy",
     type=Path(),
@@ -211,13 +214,20 @@ account_group.add_command(lntn_balances)
     is_flag=True,
     help="Echo password input to the terminal",
 )
-def new(key_file: str, extra_entropy: Optional[str], show_password: bool) -> None:
+def new(
+    keystore: Optional[str],
+    keyfile: Optional[str],
+    extra_entropy: Optional[str],
+    show_password: bool,
+) -> None:
     """
-    Create a new key and write it to a keyfile.
+    Create a new key and write it to a keyfile.  If no keyfile is
+    specified, a default name is used (consistent with GETH keyfiles)
+    in the keystore.
     """
 
     from autcli.logging import log
-    from autcli.utils import prompt_for_new_password
+    from autcli.utils import prompt_for_new_password, new_keyfile_from_options
 
     from autonity.utils.keyfile import (
         create_keyfile_from_private_key,
@@ -225,11 +235,7 @@ def new(key_file: str, extra_entropy: Optional[str], show_password: bool) -> Non
     )
 
     import json
-    import os.path
     import eth_account
-
-    if os.path.exists(key_file):
-        raise ClickException("refusing to overwrite existing keyfile")
 
     # Ask for extra entropy, if requested.
 
@@ -255,10 +261,13 @@ def new(key_file: str, extra_entropy: Optional[str], show_password: bool) -> Non
             f"internal error (address-mismatch) {account.address} != {keyfile_addr}"
         )
 
-    with open(key_file, "w", encoding="utf8") as key_f:
+    # If keyfile was not given, generate a new keyfile based on
+    # keystore and the new key details.
+    keyfile = new_keyfile_from_options(keystore, keyfile, keyfile_addr)
+    with open(keyfile, "w", encoding="utf8") as key_f:
         json.dump(keyfile_data, key_f)
 
-    log(f"Encrypted key written to {key_file}")
+    log(f"Encrypted key written to {keyfile}")
 
     print(keyfile_addr)
 
@@ -267,7 +276,8 @@ account_group.add_command(new)
 
 
 @command()
-@keyfile_option(required=True, output=True)
+@keystore_option()
+@keyfile_option(output=True)
 @option(
     "--show-password",
     is_flag=True,
@@ -275,15 +285,23 @@ account_group.add_command(new)
 )
 @argument("private_key_file", type=Path(exists=True))
 def import_private_key(
-    key_file: str, show_password: bool, private_key_file: str
+    keystore: Optional[str],
+    keyfile: Optional[str],
+    show_password: bool,
+    private_key_file: str,
 ) -> None:
     """
     Read a plaintext private key file (as hex), and create a new
     encrypted keystore file for it.  Use - to read private key from
-    stdin.
+    stdin.  If no keyfile is specified, a default name is used
+    (consistent with GETH keyfiles) in the keystore.
     """
 
-    from autcli.utils import load_from_file_or_stdin, prompt_for_new_password
+    from autcli.utils import (
+        load_from_file_or_stdin,
+        prompt_for_new_password,
+        new_keyfile_from_options,
+    )
     from autcli.logging import log
 
     from autonity.utils.keyfile import (
@@ -294,10 +312,6 @@ def import_private_key(
 
     from hexbytes import HexBytes
     import json
-    import os.path
-
-    if os.path.exists(key_file):
-        raise ClickException("refusing to overwrite existing keyfile")
 
     private_key = HexBytes.fromhex(load_from_file_or_stdin(private_key_file))
     if len(private_key) != 32:
@@ -306,11 +320,13 @@ def import_private_key(
     password = prompt_for_new_password(show_password)
 
     keyfile_data = create_keyfile_from_private_key(PrivateKey(private_key), password)
+    keyfile_addr = get_address_from_keyfile(keyfile_data)
 
-    with open(key_file, "w", encoding="utf8") as key_f:
+    keyfile = new_keyfile_from_options(keystore, keyfile, keyfile_addr)
+    with open(keyfile, "w", encoding="utf8") as key_f:
         json.dump(keyfile_data, key_f)
 
-    log(f"Encrypted key written to {key_file}")
+    log(f"Encrypted key written to {keyfile}")
 
     print(get_address_from_keyfile(keyfile_data))
 
@@ -325,7 +341,7 @@ account_group.add_command(import_private_key)
     type=Path(),
     required=True,
 )
-def signtx(key_file: Optional[str], password: Optional[str], tx_file: str) -> None:
+def signtx(keyfile: Optional[str], password: Optional[str], tx_file: str) -> None:
     """
     Sign a transaction using the given keyfile.  Use '-' to read from
     stdin instead of a file.
@@ -346,9 +362,9 @@ def signtx(key_file: Optional[str], password: Optional[str], tx_file: str) -> No
     tx = json.loads(load_from_file_or_stdin(tx_file))
 
     # Read keyfile
-    key_file = config.get_keyfile(key_file)
-    log(f"using key file: {key_file}")
-    with open(key_file, encoding="ascii") as key_f:
+    keyfile = config.get_keyfile(keyfile)
+    log(f"using key file: {keyfile}")
+    with open(keyfile, encoding="ascii") as key_f:
         encrypted_key = json.load(key_f)
 
     # Read password
@@ -377,7 +393,7 @@ account_group.add_command(signtx)
 )
 @argument("signature-file", type=Path(), required=False)
 def sign_message(
-    key_file: Optional[str],
+    keyfile: Optional[str],
     password: Optional[str],
     use_message_file: bool,
     message: str,
@@ -405,9 +421,9 @@ def sign_message(
         message = load_from_file_or_stdin(message)
 
     # Read keyfile
-    key_file = config.get_keyfile(key_file)
-    log(f"using key file: {key_file}")
-    with open(key_file, encoding="ascii") as key_f:
+    keyfile = config.get_keyfile(keyfile)
+    log(f"using key file: {keyfile}")
+    with open(keyfile, encoding="ascii") as key_f:
         encrypted_key = json.load(key_f)
 
     # Read password
@@ -451,7 +467,7 @@ account_group.add_command(sign_message)
     required=True,
 )
 def verify_signature(
-    key_file: Optional[str],
+    keyfile: Optional[str],
     from_str: Optional[str],
     use_message_file: bool,
     message: str,
@@ -482,7 +498,7 @@ def verify_signature(
         signature_hex = signature_f.read().rstrip()
         signature = HexBytes(signature_hex)
 
-    from_addr = from_address_from_argument_optional(from_str, key_file)
+    from_addr = from_address_from_argument_optional(from_str, keyfile)
 
     recovered_addr = Account().recover_message(
         encode_defunct(text=message), signature=signature