+++ autcli/utils.py
@@ -18,6 +18,7 @@ from autonity.utils.tx import (
 import os
 import sys
 import json
+from datetime import datetime, timezone
 from decimal import Decimal
 from click import ClickException
 from getpass import getpass
@@ -73,7 +74,7 @@ def autonity_from_endpoint_arg(endpoint_arg: Optional[str]) -> Autonity:
 
 
 def from_address_from_argument_optional(
-    from_str: Optional[str], key_file: Optional[str]
+    from_str: Optional[str], keyfile: Optional[str]
 ) -> Optional[ChecksumAddress]:
     """
     Given an optional command line parameter, create an address,
@@ -87,11 +88,11 @@ def from_address_from_argument_optional(
         from_addr: Optional[ChecksumAddress] = Web3.toChecksumAddress(from_str)
     else:
         log("no from-addr given.  attempting to extract from keyfile")
-        key_file = config.get_keyfile_optional(key_file)
-        if key_file:
-            key_data = load_keyfile(key_file)
+        keyfile = config.get_keyfile_optional(keyfile)
+        if keyfile:
+            key_data = load_keyfile(keyfile)
             from_addr = get_address_from_keyfile(key_data)
-            log(f"got keyfile: {key_file}, address: {from_addr}")
+            log(f"got keyfile: {keyfile}, address: {from_addr}")
         else:
             log("no keyfile.  empty from-addr")
             from_addr = None
@@ -100,14 +101,14 @@ def from_address_from_argument_optional(
 
 
 def from_address_from_argument(
-    from_str: Optional[str], key_file: Optional[str]
+    from_str: Optional[str], keyfile: Optional[str]
 ) -> ChecksumAddress:
     """
     Given an optional command line parameter, create an address,
     falling back to the keyfile given in the config.  Throws a
     ClickException if the address cannot be determined.
     """
-    from_addr = from_address_from_argument_optional(from_str, key_file)
+    from_addr = from_address_from_argument_optional(from_str, keyfile)
     if from_addr:
         return from_addr
 
@@ -477,3 +478,35 @@ def prompt_for_new_password(show_password: bool) -> str:
         raise ClickException("passwords do not match")
 
     return password
+
+
+def geth_keyfile_name(key_time: datetime, address: ChecksumAddress) -> str:
+    """
+    Given a datetime and an address, construct the base of the file
+    name of the keystore file, as used by geth.
+    """
+    # Convert the key_time into the correct format.
+    keyfile_time = key_time.strftime("%Y-%m-%dT%H-%M-%S.%f000Z")
+    # 0xca57....72EC -> ca57....72ec
+    keyfile_address = address.lower()[2:]
+    return f"UTC--{keyfile_time}--{keyfile_address}"
+
+
+def new_keyfile_from_options(
+    keystore: Optional[str], keyfile: Optional[str], keyfile_addr: ChecksumAddress
+) -> str:
+    """
+    Logic to determine a (new) keyfile name, given keystore and
+    keyfile options, where we fallback to filenames compatible with
+    geth in the keystore.  Also checks for existence of the keyfile.
+    """
+
+    if keyfile is None:
+        key_time = datetime.now(timezone.utc)
+        keystore = config.get_keystore_directory(keystore)
+        keyfile = os.path.join(keystore, geth_keyfile_name(key_time, keyfile_addr))
+
+    if os.path.exists(keyfile):
+        raise ClickException(f"refusing to overwrite existing keyfile {keyfile}")
+
+    return keyfile